web/requirements.txt
alembic==0.8.6
defusedxml==0.4.1
Flask==0.10.1
Flask-Login==0.2.11
Flask-Migrate==1.6.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
oauthlib==1.1.2
pkg-resources==0.0.0
psycopg2==2.6.1
PyJWT==1.4.0
python-editor==1.0
python-social-auth==0.2.19
python3-openid==3.0.10
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
SQLAlchemy==1.1.0b2
Werkzeug==0.11.10
WTForms==2.1
